+++ src/Resources/AdminUI.fxml
@@ -1,12 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.SplitPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.ToolBar?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="722.0" prefWidth="1174.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Controller">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="722.0" prefWidth="1174.0" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="AdminPanel.Controller">
     <bottom>
         <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
             <items>
@@ -26,7 +33,7 @@
                         <Label layoutX="26.0" layoutY="78.0" text="Room Description" />
                         <TextField fx:id="nameField" layoutX="134.0" layoutY="33.0" promptText="Enter room name" />
                         <TextField fx:id="descriptField" layoutX="134.0" layoutY="74.0" promptText="Enter Description" />
-                  <Button fx:id="addRoomBtn" layoutX="174.0" layoutY="193.0" mnemonicParsing="false" onMouseClicked="#addRoomClicked" text="Add room" />
+                  <Button fx:id="addRoomBtn" layoutX="174.0" layoutY="193.0" mnemonicParsing="false" onMouseClicked="#addRoomBtnClicked" text="Add room" />
                   <TextField layoutX="135.0" layoutY="111.0" promptText="Enter X axis" fx:id="xCoordField" />
                   <TextField fx:id="YCoordField" layoutX="135.0" layoutY="150.0" promptText="Enter Y axis" />
                   <Label layoutX="26.0" layoutY="115.0" text="Room X" />
@@ -37,7 +44,7 @@
                            <children>
                               <ImageView fitHeight="637.0" fitWidth="773.0" pickOnBounds="true" preserveRatio="true">
                                  <image>
-                                    <Image url="@../../../Image-Gallery/4_thefourthfloor.png" />
+                                    <Image url="@4_thefourthfloor.png" />
                                  </image>
                               </ImageView>
                            </children>
