+++ src/Resources/sql.d/DatabaseSchema.sql
