--- src/UI/Navigator/src/AdminController.java
@@ -1,5 +1,0 @@
-/**
- * Created by s7sal on 3/28/2017.
- */
-public class AdminController {
-}
