--- src/UI/Navigator/src/AdminUI.fxml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="722.0" prefWidth="1174.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-    <bottom>
-        <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
-            <items>
-                <Button mnemonicParsing="false" text="Logout" />
-                <Pane maxWidth="-Infinity" prefHeight="30.0" prefWidth="439.0" HBox.hgrow="ALWAYS" />
-                <Button alignment="CENTER" contentDisplay="CENTER" defaultButton="true" layoutX="150.0" mnemonicParsing="false" prefHeight="25.0" prefWidth="197.0" text="Save &amp; Apply" textAlignment="CENTER" />
-                <Pane prefHeight="30.0" prefWidth="354.0" />
-            </items>
-        </ToolBar>
-    </bottom>
-    <center>
-        <SplitPane orientation="VERTICAL" prefHeight="642.0" prefWidth="907.0" BorderPane.alignment="CENTER">
-            <items>
-                <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="499.0" prefWidth="785.0">
-                    <children>
-                        <ImageView fitHeight="540.0" fitWidth="841.0" layoutX="333.0" pickOnBounds="true" preserveRatio="true" />
-                        <Label layoutX="26.0" layoutY="37.0" text="Room Name " />
-                        <Label layoutX="26.0" layoutY="78.0" text="Room Description" />
-                        <TextField layoutX="134.0" layoutY="33.0" />
-                        <TextField layoutX="134.0" layoutY="74.0" />
-                  <Button layoutX="174.0" layoutY="193.0" mnemonicParsing="false" text="Add room" />
-                  <TextField layoutX="135.0" layoutY="111.0" />
-                  <TextField layoutX="135.0" layoutY="150.0" />
-                  <Label layoutX="26.0" layoutY="115.0" text="Room X" />
-                  <Label layoutX="26.0" layoutY="154.0" text="Room Y" />
-                    </children>
-                </AnchorPane>
-            </items>
-        </SplitPane>
-    </center>
-    <top>
-        <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER" />
-    </top>
-</BorderPane>
