+++ src/UserPanel/Controller.java
@@ -1,9 +1,11 @@
+package UserPanel;
+
 import javafx.fxml.FXML;
 import javafx.scene.Scene;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
+
 import java.awt.*;
 
 public class Controller {
@@ -14,13 +16,10 @@ public class Controller {
 
     @FXML
     private void logAsAdminClicked() throws Exception{
-        Parent root = (AnchorPane)FXMLLoader.load(getClass().getResource("AdminUI.fxml"));
-        //adminStage.setTitle("Faulkner Hospital Navigator");
-
-        primaryStage.setScene(new Scene(root, 1174, 722));
-        primaryStage.setScene(adminStage);
-        adminScene.show();
-
+      //  Parent root = (AnchorPane)FXMLLoader.load(getClass().getResource("../Resources/AdminUI.fxml"));
+     //   adminStage.setTitle("Faulkner Hospital Navigator");
+      //  primaryStage.setScene(new Scene(root, 1174, 722));
+     //   primaryStage.setScene(adminStage);
+    //    adminScene.show();
     }
-
 }
