+++ src/UserPanel/Main.java
@@ -1,8 +1,9 @@
+package UserPanel;
+
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
@@ -12,11 +13,12 @@ public class Main extends Application {
     public void start(Stage primaryStage) throws Exception{
 
 
-        Parent root = (BorderPane)FXMLLoader.load(getClass().getResource("FinalUI.fxml"));
+        Parent root = (BorderPane)FXMLLoader.load(getClass().getResource("../Resources/FinalUI.fxml"));
         primaryStage.setTitle("Faulkner Hospital Navigator");
         primaryStage.setScene(new Scene(root, 1174, 722));
         primaryStage.show();
     }
+
 
 
     public static void main(String[] args) {
